requirements.txt
Django==4.0
Pillow==10.0.0
django-filter==21.1